OBIS occurrences, South Atlantic full regional sample

Simple dataset of species and location for example purposes from 2022-04-04.

Usage

occ_SAtlantic

Format

A data frame with 1,014,006 rows and 4 variables:

decimalLongitude: unique x

decimalLatitude: unique y

species: unique species

records: number of records based on unique values of x, y, species

Source

https://obis.org/data/access